+++ crates/openapi/Cargo.toml
@@ -16,7 +16,6 @@ common_utils = { version = "0.1.0", path = "../common_utils", features = ["logs"
 common_types = { version = "0.1.0", path = "../common_types" }
 router_env = { version = "0.1.0", path = "../router_env" }
 euclid = { version = "0.1.0", path = "../euclid" }
-router = { version = "0.2.0", path = "../router"}
 
 [features]
 v2 = ["api_models/v2", "common_utils/v2", "api_models/tokenization_v2"]
