+++ crates/router/src/core/payments/routing/utils.rs
@@ -3,7 +3,9 @@ use std::collections::{HashMap, HashSet};
 use api_models::{
     open_router as or_types,
     routing::{
-        self as api_routing, ConnectorSelection, ConnectorVolumeSplit, RoutableConnectorChoice,
+        self as api_routing, ComparisonType, ConnectorSelection, ConnectorVolumeSplit,
+        DeRoutableConnectorChoice, MetadataValue, NumberComparison, RoutableConnectorChoice,
+        RoutingEvaluateRequest, RoutingEvaluateResponse, ValueType,
     },
 };
 use async_trait::async_trait;
@@ -24,7 +26,6 @@ use hyperswitch_domain_models::business_profile;
 use hyperswitch_interfaces::events::routing_api_logs as routing_events;
 use router_env::tracing_actix_web::RequestId;
 use serde::{Deserialize, Serialize};
-use utoipa::ToSchema;
 
 use super::RoutingResult;
 use crate::{
@@ -684,100 +685,6 @@ impl DecisionEngineErrorsInterface for or_types::ErrorResponse {
     }
 }
 
-//TODO: temporary change will be refactored afterwards
-#[derive(Clone, Debug, serde::Serialize, serde::Deserialize, PartialEq, ToSchema)]
-pub struct RoutingEvaluateRequest {
-    pub created_by: String,
-    #[schema(value_type = Object)]
-    pub parameters: HashMap<String, Option<ValueType>>,
-    pub fallback_output: Vec<DeRoutableConnectorChoice>,
-}
-impl common_utils::events::ApiEventMetric for RoutingEvaluateRequest {}
-
-#[derive(Debug, serde::Serialize, serde::Deserialize, Clone, ToSchema)]
-pub struct RoutingEvaluateResponse {
-    pub status: String,
-    pub output: serde_json::Value,
-    #[serde(deserialize_with = "deserialize_connector_choices")]
-    pub evaluated_output: Vec<RoutableConnectorChoice>,
-    #[serde(deserialize_with = "deserialize_connector_choices")]
-    pub eligible_connectors: Vec<RoutableConnectorChoice>,
-}
-impl common_utils::events::ApiEventMetric for RoutingEvaluateResponse {}
-/// Routable Connector chosen for a payment
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize, ToSchema)]
-pub struct DeRoutableConnectorChoice {
-    pub gateway_name: common_enums::RoutableConnectors,
-    pub gateway_id: Option<id_type::MerchantConnectorAccountId>,
-}
-
-fn deserialize_connector_choices<'de, D>(
-    deserializer: D,
-) -> Result<Vec<RoutableConnectorChoice>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let infos = Vec::<DeRoutableConnectorChoice>::deserialize(deserializer)?;
-    Ok(infos
-        .into_iter()
-        .map(RoutableConnectorChoice::from)
-        .collect())
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema)]
-pub struct MetadataValue {
-    pub key: String,
-    pub value: String,
-}
-
-/// Represents a value in the DSL
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema)]
-#[serde(tag = "type", content = "value", rename_all = "snake_case")]
-pub enum ValueType {
-    /// Represents a number literal
-    Number(u64),
-    /// Represents an enum variant
-    EnumVariant(String),
-    /// Represents a Metadata variant
-    MetadataVariant(MetadataValue),
-    /// Represents a arbitrary String value
-    StrValue(String),
-    /// Represents a global reference, which is a reference to a global variable
-    GlobalRef(String),
-    /// Represents an array of numbers. This is basically used for
-    /// "one of the given numbers" operations
-    /// eg: payment.method.amount = (1, 2, 3)
-    NumberArray(Vec<u64>),
-    /// Similar to NumberArray but for enum variants
-    /// eg: payment.method.cardtype = (debit, credit)
-    EnumVariantArray(Vec<String>),
-    /// Like a number array but can include comparisons. Useful for
-    /// conditions like "500 < amount < 1000"
-    /// eg: payment.amount = (> 500, < 1000)
-    NumberComparisonArray(Vec<NumberComparison>),
-}
-
-pub type Metadata = HashMap<String, serde_json::Value>;
-/// Represents a number comparison for "NumberComparisonArrayValue"
-#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct NumberComparison {
-    pub comparison_type: ComparisonType,
-    pub number: u64,
-}
-
-/// Conditional comparison type
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum ComparisonType {
-    Equal,
-    NotEqual,
-    LessThan,
-    LessThanEqual,
-    GreaterThan,
-    GreaterThanEqual,
-}
-
 /// Represents a single comparison condition.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
@@ -877,16 +784,6 @@ impl ConnectorInfo {
     }
 }
 
-impl From<DeRoutableConnectorChoice> for RoutableConnectorChoice {
-    fn from(choice: DeRoutableConnectorChoice) -> Self {
-        Self {
-            choice_kind: api_routing::RoutableChoiceKind::FullStruct,
-            connector: choice.gateway_name,
-            merchant_connector_id: choice.gateway_id,
-        }
-    }
-}
-
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum Output {
@@ -898,6 +795,7 @@ pub enum Output {
 
 pub type Globals = HashMap<String, HashSet<ValueType>>;
 
+pub type Metadata = HashMap<String, serde_json::Value>;
 /// The program, having a default connector selection and
 /// a bunch of rules. Also can hold arbitrary metadata.
 #[derive(Clone, Debug, Serialize, Deserialize)]
